+++ packages/bytedance/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/bytedance
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [58bc42d]
+  - @ai-sdk/provider-utils@4.0.16
+
 ## 1.0.0
 
 ### Major Changes