+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/gateway
 
+## 3.0.60
+
+### Patch Changes
+
+- 29e9f4d: chore(provider/gateway): update gateway model settings files
+
+## 3.0.59
+
+### Patch Changes
+
+- Updated dependencies [58bc42d]
+  - @ai-sdk/provider-utils@4.0.16
+
 ## 3.0.58
 
 ### Patch Changes