+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/groq
 
+## 3.0.26
+
+### Patch Changes
+
+- Updated dependencies [58bc42d]
+  - @ai-sdk/provider-utils@4.0.16
+
 ## 3.0.25
 
 ### Patch Changes