+++ packages/prodia/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/prodia
 
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [58bc42d]
+  - @ai-sdk/provider-utils@4.0.16
+
 ## 1.0.17
 
 ### Patch Changes